#174b Move HighscoreTests from `level.save` to `server`
<organization
