+++ src/test/java/nl/tudelft/ti2206/group9/server/HighscoreTest.java
@@ -1,9 +1,8 @@
-package nl.tudelft.ti2206.group9.level.save;
+package nl.tudelft.ti2206.group9.server;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
-import nl.tudelft.ti2206.group9.server.Highscore;
 
 import org.junit.Test;
